requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-genai>=0.1.0
supabase>=2.0.0
pinecone>=6.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# langgraph 1.x split prebuilt into a separate package; both are required
langgraph>=1.0.0
langgraph-prebuilt>=1.0.0
# 3.1.0+ preserves thought_signature for Gemini 3.x multi-turn tool calls (required by API)
langchain-google-genai>=3.1.0
sse-starlette>=2.0.0
